GH-488: chains strongly asserted; next I'm going to remove Rinkeby as decided

+++ masq_lib/src/test_utils/utils.rs
@@ -3,8 +3,8 @@
 use std::fs;
 use std::path::PathBuf;
 
-pub const TEST_DEFAULT_CHAIN_ID: u8 = 3;
-pub const TEST_DEFAULT_MULTINODE_CHAIN_ID: u8 = 2;
+pub const TEST_DEFAULT_CHAIN_ID: u64 = 3;
+pub const TEST_DEFAULT_MULTINODE_CHAIN_ID: u64 = 2;
 pub const TEST_DEFAULT_CHAIN_NAME: &str = "ropsten";
 pub const TEST_DEFAULT_PLATFORM: &str = "eth";
 pub const TEST_DEFAULT_MULTINODE_CHAIN_NAME: &str = "dev";
