+++ node/src/sub_lib/blockchain_bridge.rs
@@ -13,7 +13,7 @@ use std::fmt::{Debug, Formatter};
 #[derive(Clone, PartialEq, Debug, Default)]
 pub struct BlockchainBridgeConfig {
     pub blockchain_service_url_opt: Option<String>,
-    pub chain_id: u8,
+    pub chain_id: u64,
     pub gas_price: u64,
 }
